+++ src/main/java/com/tallerwebi/dominio/servicios/ServicioResumenInteligente.java
@@ -1,4 +1,4 @@
-package com.tallerwebi.dominio;
+package com.tallerwebi.dominio.servicios;
 
 import com.tallerwebi.dominio.DTO.MateriaDTO;
 import org.springframework.http.*;
