+++ src/main/java/com/tallerwebi/infraestructura/RepositorioCarreraImpl.java
@@ -1,13 +1,12 @@
 package com.tallerwebi.infraestructura;
 
 import com.tallerwebi.dominio.Carrera;
-import com.tallerwebi.dominio.RepositorioCarrera;
+import com.tallerwebi.repositorioInterfaz.RepositorioCarrera;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-import javax.transaction.Transactional;
 import java.util.List;
 
 @Repository("repositorioCarrera")
