+++ src/main/java/com/tallerwebi/infraestructura/RepositorioComentarioImpl.java
@@ -1,7 +1,7 @@
 package com.tallerwebi.infraestructura;
 
 import com.tallerwebi.dominio.Comentario;
-import com.tallerwebi.dominio.RepositorioComentario;
+import com.tallerwebi.repositorioInterfaz.RepositorioComentario;
 import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
