+++ src/main/java/com/tallerwebi/infraestructura/RepositorioLoginImplementacion.java
@@ -1,7 +1,7 @@
 package com.tallerwebi.infraestructura;
 
-import com.tallerwebi.dominio.RepositorioLogin;
-import com.tallerwebi.dominio.RepositorioUsuario;
+import com.tallerwebi.repositorioInterfaz.RepositorioLogin;
+import com.tallerwebi.repositorioInterfaz.RepositorioUsuario;
 import com.tallerwebi.dominio.Usuario;
 import com.tallerwebi.dominio.excepcion.UsuarioExistente;
 import org.springframework.beans.factory.annotation.Autowired;
