+++ src/main/java/com/tallerwebi/presentacion/ControladorEmail.java
@@ -1,7 +1,7 @@
 package com.tallerwebi.presentacion;
 
 import com.tallerwebi.dominio.Usuario;
-import com.tallerwebi.dominio.ServicioEmail;
+import com.tallerwebi.dominio.servicios.ServicioEmail;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
