+++ src/main/java/com/tallerwebi/presentacion/ControladorForo.java
@@ -2,14 +2,16 @@ package com.tallerwebi.presentacion;
 
 import com.tallerwebi.dominio.*;
 import com.tallerwebi.dominio.excepcion.UsuarioNoEncontrado;
+import com.tallerwebi.dominio.servicios.ServicioMateria;
+import com.tallerwebi.servicioInterfaz.ServicioComentario;
+import com.tallerwebi.servicioInterfaz.ServicioPublicacion;
+import com.tallerwebi.servicioInterfaz.ServicioUsuario;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import javax.servlet.http.HttpSession;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 
 @Controller
 @RequestMapping("/foro")
