+++ src/main/java/com/tallerwebi/repositorioInterfaz/RepositorioLogin.java
@@ -1,5 +1,6 @@
-package com.tallerwebi.dominio;
+package com.tallerwebi.repositorioInterfaz;
 
+import com.tallerwebi.dominio.Usuario;
 import com.tallerwebi.dominio.excepcion.UsuarioExistente;
 
 public interface RepositorioLogin {
