+++ src/main/java/com/tallerwebi/repositorioInterfaz/RepositorioMateria.java
@@ -1,4 +1,7 @@
-package com.tallerwebi.dominio;
+package com.tallerwebi.repositorioInterfaz;
+
+import com.tallerwebi.dominio.Carrera;
+import com.tallerwebi.dominio.Materia;
 
 import java.util.List;
