+++ src/main/java/com/tallerwebi/repositorioInterfaz/RepositorioUsuario.java
@@ -1,4 +1,6 @@
-package com.tallerwebi.dominio;
+package com.tallerwebi.repositorioInterfaz;
+
+import com.tallerwebi.dominio.Usuario;
 
 public interface RepositorioUsuario {
