+++ src/main/java/com/tallerwebi/servicioInterfaz/ServicioComentario.java
@@ -1,5 +1,6 @@
-package com.tallerwebi.dominio;
+package com.tallerwebi.servicioInterfaz;
 
+import com.tallerwebi.dominio.Usuario;
 import com.tallerwebi.dominio.excepcion.AccesoDenegado;
 import com.tallerwebi.dominio.excepcion.ComentarioInexistente;
 import com.tallerwebi.dominio.excepcion.PublicacionInexistente;
