+++ src/main/java/com/tallerwebi/servicioInterfaz/ServicioUsuario.java
@@ -1,5 +1,6 @@
-package com.tallerwebi.dominio;
+package com.tallerwebi.servicioInterfaz;
 
+import com.tallerwebi.dominio.Usuario;
 import com.tallerwebi.dominio.excepcion.UsuarioNoEncontrado;
 
 public interface ServicioUsuario {
